requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==15.3.4
h11==0.14.0
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
Markdown==3.4.1
mysqlclient==2.1.1
outcome==1.2.0
packaging==22.0
Pillow==9.3.0
pluggy==1.0.0
py==1.11.0
Pygments==2.13.0
PySocks==1.7.1
pytest==6.2.5
pytest-base-url==2.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
pytest-html==3.2.0
pytest-metadata==2.0.4
pytest-selenium==4.0.0
pytest-variables==2.0.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.7
requests==2.28.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tenacity==6.3.1
toml==0.10.2
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.13
wsproto==1.2.0
